Use InteractiveCanvas2D for biome source preview
