+++ src/app/previews/index.ts
@@ -1,2 +1,1 @@
-export * from './BiomeSource.js'
 export * from './Decorator.js'
