+++ src/app/schema/renderHtml.tsx
@@ -6,10 +6,10 @@ import { memo } from 'preact/compat'
 import { useState } from 'preact/hooks'
 import { Btn, Octicon } from '../components/index.js'
 import { ItemDisplay } from '../components/ItemDisplay.jsx'
+import { VanillaColors } from '../components/previews/BiomeSourcePreview.jsx'
 import config from '../Config.js'
 import { localize, useLocale, useStore } from '../contexts/index.js'
 import { useFocus } from '../hooks/index.js'
-import { VanillaColors } from '../previews/index.js'
 import type { BlockStateRegistry, VersionId } from '../services/index.js'
 import { CachedDecorator, CachedFeature } from '../services/index.js'
 import { deepClone, deepEqual, generateUUID, hexId, hexToRgb, isObject, newSeed, rgbToHex, stringToColor } from '../Utils.js'
