+++ src/locales/en.json
@@ -69,15 +69,14 @@
   "indentation.tabs": "Tabs",
   "item_modifier": "Item Modifier",
   "language": "Language",
+  "layer": "Layer",
   "layer.biomes": "Biomes",
   "layer.temperature": "Temperature",
-  "layer.humidity": "Humidity",
-  "layer.continentalness": "Continentalness",
+  "layer.vegetation": "Humidity",
+  "layer.continents": "Continentalness",
   "layer.erosion": "Erosion",
-  "layer.weirdness": "Weirdness",
-  "layer.offset": "Offset",
-  "layer.factor": "Factor",
-  "layer.jaggedness": "Jaggedness",
+  "layer.ridges": "Weirdness",
+  "layer.depth": "Depth",
   "highlighting": "Highlighting",
   "loading": "Loading...",
   "loot_table": "Loot Table",
